Remove GenAxisArray base class. Set zero_copy=True on ModifyAxis with some tests.
