+++ src/ezmsg/util/messages/modify.py
@@ -1,11 +1,12 @@
 from dataclasses import replace
+import traceback
 import typing
 
 import numpy as np
 
 import ezmsg.core as ez
 from ezmsg.util.messages.axisarray import AxisArray
-from ezmsg.util.generator import consumer, GenAxisArray
+from ezmsg.util.generator import consumer, GenState
 
 
 @consumer
@@ -31,8 +32,33 @@ class ModifyAxisSettings(ez.Settings):
     name_map: typing.Optional[typing.Dict[str, str]] = None
 
 
-class ModifyAxis(GenAxisArray):
+class ModifyAxis(ez.Unit):
+    STATE: GenState
     SETTINGS: ModifyAxisSettings
+
+    INPUT_SIGNAL = ez.InputStream(AxisArray)
+    OUTPUT_SIGNAL = ez.OutputStream(AxisArray)
+    INPUT_SETTINGS = ez.InputStream(ez.Settings)
+
+    def initialize(self) -> None:
+        self.construct_generator()
 
     def construct_generator(self):
         self.STATE.gen = modify_axis(name_map=self.SETTINGS.name_map)
+
+    @ez.subscriber(INPUT_SETTINGS)
+    async def on_settings(self, msg: ez.Settings) -> None:
+        self.apply_settings(msg)
+        self.construct_generator()
+
+    @ez.subscriber(INPUT_SIGNAL, zero_copy=True)
+    @ez.publisher(OUTPUT_SIGNAL)
+    async def on_message(self, message: AxisArray) -> typing.AsyncGenerator:
+        try:
+            ret = self.STATE.gen.send(message)
+            if ret is not None:
+                yield self.OUTPUT_SIGNAL, ret
+        except (StopIteration, GeneratorExit):
+            ez.logger.debug(f"Generator closed in {self.address}")
+        except Exception:
+            ez.logger.info(traceback.format_exc())
