+++ tests/messages/test_modify.py
@@ -1,6 +1,8 @@
+import copy
+import typing
+
 import numpy as np
 import pytest
-import typing
 
 from ezmsg.util.messages.axisarray import AxisArray
 from ezmsg.util.messages.modify import modify_axis, ModifyAxis, ModifyAxisSettings
@@ -13,9 +15,19 @@ def test_modify_axis(name_map: typing.Optional[typing.Dict[str, str]]):
         dims=["step", "freq", "ch"],
         axes={"step": AxisArray.Axis.TimeAxis(fs=10.0, offset=0.0)},
     )
+    backup = copy.deepcopy(input_ax_arr)
 
     gen = modify_axis(name_map)
     res = gen.send(input_ax_arr)
+
+    # Make sure the input hasn't changed
+    assert np.array_equal(input_ax_arr.data, backup.data)
+    assert input_ax_arr.dims == backup.dims
+    assert list(input_ax_arr.axes.keys()) == list(backup.axes.keys())
+    for k, v in input_ax_arr.axes.items():
+        assert v == backup.axes[k]
+
+    assert res.data is input_ax_arr.data
     if name_map is None:
         assert res is input_ax_arr
     else:
